+++ src/main/container/type.ts
@@ -1,3 +1,4 @@
+import { ICreateTodoService, IListService } from '../../core/services';
 import {
   IJsonPlaceHolderIntegration,
   ITodoCache,
@@ -8,4 +9,9 @@ export type InfraContext = {
   todo_repository: ITodoRepository;
   todo_cache: ITodoCache;
   json_place_holder_integration: IJsonPlaceHolderIntegration;
+};
+
+export type UseCaseContext = {
+  create_todo_service: ICreateTodoService;
+  list_todo_service: IListService;
 };