+++ CHANGELOG.md
@@ -7,10 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Ongoing Changes
 
+- ref - create service layer
 - add - Readme
 - add - Dockerfile
 - add - env validation
-- add - cache 
+- add - cache
 - add - db and migrations to database
 - fix - adjust errors for cli interface
 - fix - add port in constructor of http module